refactoring client credientals authentication
